+++ src/main/java/model/lotto/LottoCount.java
@@ -1,4 +1,4 @@
-package model;
+package model.lotto;
 
 import utils.InputValidateUtils;
