+++ src/main/java/model/lotto/LottoDTO.java
@@ -1,4 +1,4 @@
-package model;
+package model.lotto;
 
 import java.util.ArrayList;
 import java.util.List;
