+++ src/main/java/model/winningnumber/LottoWinningNumber.java
@@ -1,9 +1,7 @@
-package model;
+package model.winningnumber;
 
 import utils.InputValidateUtils;
 
-import java.sql.Array;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
