+++ src/test/java/model/bonusball/BonusBallTest.java
@@ -1,8 +1,9 @@
-package model;
+package model.bonusball;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+import model.bonusball.BonusBall;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
