+++ src/test/java/model/lotto/LottoCountTest.java
@@ -1,8 +1,9 @@
-package model;
+package model.lotto;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+import model.lotto.LottoCount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
