+++ src/test/java/model/lotto/LottoStorageTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.lotto;
 
 public class LottoStorageTest {
 }
