+++ src/test/java/model/lotto/LottoTest.java
@@ -1,7 +1,11 @@
-package model;
+package model.lotto;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
+import model.bonusball.BonusBallDTO;
+import model.lotto.Lotto;
+import model.result.Statistics;
+import model.winningnumber.LottoWinningNumberDTO;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
