+++ src/test/java/model/lotto/RandomLottoNumbersGeneratorTest.java
@@ -1,7 +1,8 @@
-package model;
+package model.lotto;
 
 import java.util.List;
 
+import model.lotto.RandomLottoNumbersGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
