+++ src/test/java/model/winningnumber/LottoWinningNumberTest.java
@@ -1,8 +1,9 @@
-package model;
+package model.winningnumber;
 
 import java.util.Arrays;
 import java.util.List;
 
+import model.winningnumber.LottoWinningNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
